+++ arch/sim/src/Makefile
@@ -61,7 +61,7 @@ CSRCS  = up_initialize.c up_idle.c up_interruptcontext.c up_initialstate.c
 CSRCS += up_createstack.c up_usestack.c up_releasestack.c up_stackframe.c
 CSRCS += up_unblocktask.c up_blocktask.c up_releasepending.c
 CSRCS += up_reprioritizertr.c up_exit.c up_schedulesigaction.c
-CSRCS += up_heap.c up_uart.c
+CSRCS += up_heap.c up_uart.c up_assert.c
 CSRCS += up_copyfullstate.c
 
 ifeq ($(CONFIG_ARCH_HAVE_VFORK),y)
